feat: separated the post card component to use it in the category page also
Close: #29
